pyproject.toml
[tool.poetry]
name = "silver-flag-bot"
authors = ["HM <hieumdd@gmail.com>"]
package-mode = false

[tool.poetry.dependencies]
python = "~3.11"
python-json-logger = "^2.0.7"
httpx = ">=0.25,<1.0"
pyjwt = "^2.8.0"
cryptography = "^42.0.0"
python-telegram-bot = { extras = ["job-queue"], version = "^20.7" }
pandas = "^2.2.0"
pandas-ta = "^0.3.14b0"
emoji = "^2.10.0"
apscheduler = "^3.10.4"
matplotlib = "^3.8.2"
mplfinance = "^0.12.10b0"
yattag = "^1.15.2"
pyyaml = "^6.0.1"


[tool.poetry.group.dev.dependencies]
black = "^23.11.0"
mypy = "^1.7.0"
python-dotenv = { extras = ["cli"], version = "^1.0.1" }
pytest = "^7.4.4"
pytest-dotenv = "^0.5.2"
notebook = "^7.0.7"
pandas-stubs = "^2.1.4.231227"

[tool.mypy]
explicit_package_bases = true

[tool.pytest.ini_options]
python_files = "*_test.py"
env_files = ".env"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
